requirements.txt
agate==1.9.1
annotated-types==0.7.0
asn1crypto==1.5.1
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
-e git+ssh://git@github.com/Sanourith/BTC_app.git@74e58cd82cd543d6208dc783b0f0b02fc043c189#egg=btc_functions
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
daff==1.3.46
dbt==1.0.0.38.22
dbt-adapters==1.10.3
dbt-common==1.13.0
dbt-core==1.8.9
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1
dbt-snowflake==1.8.4
debugpy==1.8.8
decorator==5.1.1
deepdiff==7.0.1
executing==2.1.0
filelock==3.16.1
fonttools==4.55.0
greenlet==3.1.1
idna==3.10
importlib-metadata==6.11.0
ipykernel==6.29.5
ipython==8.29.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.5.0
kiwisolver==1.4.7
leather==0.4.0
Logbook==1.5.3
MarkupSafe==3.0.2
mashumaro==3.14
matplotlib==3.9.2
matplotlib-inline==0.1.7
minimal-snowplow-tracker==0.0.2
more-itertools==10.5.0
msgpack==1.1.0
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.3
setuptools==75.6.0
six==1.16.0
snowflake-connector-python==3.12.3
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlparse==0.5.2
stack-data==0.6.3
text-unidecode==1.3
threadpoolctl==3.5.0
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
unzip==1.0.0
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.21.0
